requirements.txt
# Core Libraries
numpy>=1.21.0
pandas>=1.3.3
kaggle>=1.6.3
# Data Visualization
matplotlib>=3.4.3
seaborn>=0.13.0
pillow>=9.3.0

tqdm>=4.66.0

# Web Application Framework
streamlit>=1.28.0

scikit-learn>=0.24.2
torch==2.1.2
torchvision>=0.16.2
albumentations>=1.3.1